requirements.txt
streamlit
cohere
PyPDF2
pdfplumber
python-docx
reportlab
